refactor(command-registry): move the batch runner and batch policy into @agent-device/command-registry
