+++ packages/command-registry/package.json
@@ -7,6 +7,7 @@
   "description": "The command descriptor registry: the single declaration of every command's surface, runtime-use, batch, timeout, and platform-execution traits, plus the catalogs and projections derived from it.",
   "dependencies": {
     "@agent-device/contracts": "workspace:*",
+    "@agent-device/kernel": "workspace:*",
     "@agent-device/selectors": "workspace:*"
   },
   "exports": {
@@ -61,6 +62,14 @@
     "./wait-positionals": {
       "types": "./src/wait-positionals.ts",
       "default": "./src/wait-positionals.ts"
+    },
+    "./batch": {
+      "types": "./src/batch.ts",
+      "default": "./src/batch.ts"
+    },
+    "./batch-policy": {
+      "types": "./src/batch-policy.ts",
+      "default": "./src/batch-policy.ts"
     }
   }
 }
