+++ packages/command-registry/src/__tests__/batch.test.ts
