+++ packages/command-registry/src/batch.ts
