+++ packages/contracts/src/batch-step.ts
@@ -3,9 +3,11 @@ import type { SessionRuntimeHints } from '@agent-device/kernel/contracts';
 /**
  * One step of a daemon batch, as submitted.
  *
- * Declared here rather than in `core/batch.ts` because the public API vocabulary
- * (`contracts/client-replay.ts`) is stated in terms of it, and `core/` sits above contracts. The
- * `runtime` field used to be written as `DaemonRequest['runtime']`, which pulled the whole daemon
+ * Declared here rather than in `@agent-device/command-registry/batch` because the public API
+ * vocabulary (`contracts/client-replay.ts`) is stated in terms of it, and contracts sits below
+ * command-registry.
+ *
+ * The `runtime` field used to be written as `DaemonRequest['runtime']`, which pulled the whole daemon
  * request type in to say `SessionRuntimeHints` — the same type, one zone lower.
  */
 export type DaemonBatchStep = {
